+++ airline-project/src/main/java/app/exceptions/BusinessException.java
@@ -1,0 +1,16 @@
+package app.exceptions;
+
+
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+public abstract class BusinessException extends RuntimeException {
+
+    @Getter
+    private final HttpStatus httpStatus;
+
+    protected BusinessException(String message, HttpStatus httpStatus) {
+        super(message);
+        this.httpStatus = httpStatus;
+    }
+}