+++ airline-project/src/main/java/app/exceptions/EntityNotFoundException.java
@@ -1,8 +1,9 @@
 package app.exceptions;
 
-public class EntityNotFoundException extends RuntimeException {
+import org.springframework.http.HttpStatus;
 
+public class EntityNotFoundException extends BusinessException {
     public EntityNotFoundException(String message) {
-        super(message);
+        super(message, HttpStatus.NOT_FOUND);
     }
 }